Create a new "TempDir" option for atomic write
The current atomic write is somewhat broken as the temporary files are
created inside the base directory, and so count as potential keys (and
can conflict). Move these temporary files in the TempDir specified by
the user so that they can't conflict with future keys.

Note that the TempDir must be on the same device/partition/mount or the
atomic Rename will fail across partition.

Atomic write will not happen if TempDir is left empty.

+++ diskv.go
@@ -46,6 +46,7 @@ type Options struct {
 	CacheSizeMax uint64 // bytes
 	PathPerm     os.FileMode
 	FilePerm     os.FileMode
+	TempDir      string
 
 	Index     Index
 	IndexLess LessFunction
@@ -115,21 +116,43 @@ func (d *Diskv) WriteStream(key string, r io.Reader, sync bool) error {
 	return d.writeStreamWithLock(key, r, sync)
 }
 
+// createKeyFileWithLock either creates the key file directly, or
+// creates a temporary file in TempDir if it is set.
+func (d *Diskv) createKeyFileWithLock(key string) (*os.File, error) {
+	if d.TempDir != "" {
+		if err := os.MkdirAll(d.TempDir, d.PathPerm); err != nil {
+			return nil, fmt.Errorf("temp mkdir: %s", err)
+		}
+		f, err := ioutil.TempFile(d.TempDir, "")
+		if err != nil {
+			return nil, fmt.Errorf("temp file: %s", err)
+		}
+
+		if err := f.Chmod(d.FilePerm); err != nil {
+			f.Close()           // error deliberately ignored
+			os.Remove(f.Name()) // error deliberately ignored
+			return nil, fmt.Errorf("chmod: %s", err)
+		}
+		return f, nil
+	}
+
+	mode := os.O_WRONLY | os.O_CREATE | os.O_TRUNC // overwrite if exists
+	f, err := os.OpenFile(d.completeFilename(key), mode, d.FilePerm)
+	if err != nil {
+		return nil, fmt.Errorf("open file: %s", err)
+	}
+	return f, nil
+}
+
 // writeStream does no input validation checking.
 func (d *Diskv) writeStreamWithLock(key string, r io.Reader, sync bool) error {
 	if err := d.ensurePathWithLock(key); err != nil {
 		return fmt.Errorf("ensure path: %s", err)
 	}
 
-	f, err := ioutil.TempFile(d.pathFor(key), fmt.Sprintf("temp-%s-", key))
+	f, err := d.createKeyFileWithLock(key)
 	if err != nil {
-		return fmt.Errorf("temp file: %s", err)
-	}
-
-	if err := f.Chmod(d.FilePerm); err != nil {
-		f.Close()           // error deliberately ignored
-		os.Remove(f.Name()) // error deliberately ignored
-		return fmt.Errorf("chmod: %s", err)
+		return fmt.Errorf("create key file: %s", err)
 	}
 
 	wc := io.WriteCloser(&nopWriteCloser{f})
@@ -166,9 +189,11 @@ func (d *Diskv) writeStreamWithLock(key string, r io.Reader, sync bool) error {
 		return fmt.Errorf("file close: %s", err)
 	}
 
-	if err := os.Rename(f.Name(), d.completeFilename(key)); err != nil {
-		os.Remove(f.Name()) // error deliberately ignored
-		return fmt.Errorf("rename: %s", err)
+	if f.Name() != d.completeFilename(key) {
+		if err := os.Rename(f.Name(), d.completeFilename(key)); err != nil {
+			os.Remove(f.Name()) // error deliberately ignored
+			return fmt.Errorf("rename: %s", err)
+		}
 	}
 
 	if d.Index != nil {
@@ -403,6 +428,9 @@ func (d *Diskv) EraseAll() error {
 	defer d.mu.Unlock()
 	d.cache = make(map[string][]byte)
 	d.cacheSize = 0
+	if d.TempDir != "" {
+		os.RemoveAll(d.TempDir) // errors ignored
+	}
 	return os.RemoveAll(d.BasePath)
 }
